+++ src/main/java/com/campus/trade/entity/Message.java
@@ -5,13 +5,13 @@ import java.time.LocalDateTime;
 
 @Data
 public class Message {
-    private Integer id;
-    private Integer senderId;     // 发送者ID
-    private Integer receiverId;   // 接收者ID
-    private Integer productId;    // 关联商品ID（可选，方便溯源）
+    private Long id;              // 消息ID（改为Long）
+    private Long senderId;        // 发送者ID（改为Long）
+    private Long receiverId;      // 接收者ID（改为Long）
+    private Long productId;       // 关联商品ID（改为Long）
     private String content;       // 消息内容（文字或图片URL）
-    private Integer type;          // 0-文本，1-交易卡片
-    private Integer isRead;        // 0-未读，1-已读
+    private Integer type;         // 0-文本，1-交易卡片
+    private Integer isRead;       // 0-未读，1-已读
     private LocalDateTime createTime;
     
     // 交易相关字段（用于交易卡片消息）
